remove paper db
