+++ libs/pandautils/src/main/java/com/instructure/pandautils/base/BaseCanvasBottomSheetDialogFragment.kt
@@ -16,12 +16,11 @@
 
 package com.instructure.pandautils.base
 
-import android.content.Context
 import android.os.Bundle
 import android.view.View
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.instructure.canvasapi2.utils.pageview.PageViewWindowFocus
 import com.instructure.pandautils.analytics.pageview.PageViewUtils
+import com.instructure.pandautils.analytics.pageview.PageViewWindowFocus
 import com.instructure.pandautils.utils.showMasqueradeNotification
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -47,9 +46,9 @@ open class BaseCanvasBottomSheetDialogFragment : BottomSheetDialogFragment(), Pa
         delegate.completePageViewPrerequisite(prerequisite)
     }
 
-    override fun onAttach(context: Context) {
-        delegate.onAttach(context)
-        super.onAttach(context)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        delegate.onCreate()
+        super.onCreate(savedInstanceState)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
